Sprint 16

+++ Tracker/CoreData/TrackerCategoryStore.swift
@@ -9,12 +9,21 @@ import Foundation
 import UIKit
 import CoreData
 
+
+
+
 final class TrackerCategoryStore: NSObject {
+    
+    
+    weak var delegate: TrackerCategoryStoreDelegateProtocol?
     
     // MARK: - Private Properties
     //
     private let context: NSManagedObjectContext
-
+    
+    private var fetchedResultsController: NSFetchedResultsController<TrackerCategoryCoreData>!
+    
+    
     // MARK: - Initializers
     //
     convenience override init() {
@@ -24,22 +33,27 @@ final class TrackerCategoryStore: NSObject {
     
     init(context: NSManagedObjectContext) throws {
         self.context = context
+        super.init()
+        
+        let fetchRequest = TrackerCategoryCoreData.fetchRequest()
+        fetchRequest.sortDescriptors = [
+            NSSortDescriptor(keyPath: \TrackerCategoryCoreData.categoryName, ascending: true)
+        ]
+        let controller = NSFetchedResultsController(
+            fetchRequest: fetchRequest,
+            managedObjectContext: context,
+            sectionNameKeyPath: nil,
+            cacheName: nil
+        )
+        
+        controller.delegate = self
+        self.fetchedResultsController = controller
+        try controller.performFetch()
     }
 }
 
 
 extension TrackerCategoryStore: TrackerCategoryStoreDataProviderProtocol{
-    
-    func getTrackerCategory(trackerID: UUID) -> String? {
-        var records: [TrackerCategoryCoreData] = []
-        let request = NSFetchRequest<TrackerCategoryCoreData>(entityName: "TrackerCategoryCoreData")
-        request.returnsObjectsAsFaults = false
-        request.predicate = NSPredicate(format:  "%K CONTAINS[n] %@", #keyPath(TrackerCategoryCoreData.trackerIDs), trackerID.uuidString)
-        
-        do { records = try context.fetch(request) } catch { return nil }
-        
-        return records.first?.categoryName
-    }
     
     func getCategoriesCount() -> Int {
         let request = NSFetchRequest<TrackerCategoryCoreData>(entityName: "TrackerCategoryCoreData")
@@ -51,11 +65,13 @@ extension TrackerCategoryStore: TrackerCategoryStoreDataProviderProtocol{
         return countRecords
     }
     
-    func updateCategory(_ category: TrackerCategory) -> Bool  {
+    func updateCategory(category: TrackerCategory, newCategory: TrackerCategory) -> Bool  {
+        
         guard let categoryObject = getCategoryObject(category.categoryName) else { return false }
         
-        categoryObject.trackerIDs = category.trackerIDsString
+        categoryObject.categoryName = newCategory.categoryName
         do { try context.save() } catch { return false }
+       // delegate?.didUpdate()
         return true
     }
     
@@ -63,54 +79,72 @@ extension TrackerCategoryStore: TrackerCategoryStoreDataProviderProtocol{
         let trackerCategoryCoreData = TrackerCategoryCoreData(context: context)
         
         trackerCategoryCoreData.categoryName = category.categoryName
-        trackerCategoryCoreData.trackerIDs = category.trackerIDsString
         
         do { try context.save() } catch { return false }
+        
+       // delegate?.didUpdate()
         return true
     }
     
     func getCategory(_ category: String) -> TrackerCategory? {
         guard let categoryObject = getCategoryObject(category),
-              let trackerIDs = categoryObject.trackerIDs,
               let categoryName = categoryObject.categoryName else { return nil}
         
-        return TrackerCategory(
-            trackerIDs: TrackerCategory.trackerIDsFromString(udids: trackerIDs),
-            categoryName: categoryName)
+        return TrackerCategory(categoryName: categoryName)
     }
     
     func getCategories() -> [TrackerCategory]? {
         let request = NSFetchRequest<TrackerCategoryCoreData>(entityName: "TrackerCategoryCoreData")
         request.returnsObjectsAsFaults = false
+        request.sortDescriptors = [NSSortDescriptor(key: "categoryName", ascending: true)]
         
         var records:[TrackerCategoryCoreData] = []
         do { records = try context.fetch(request) } catch { return nil }
         
         return records.compactMap {
-            guard let trackerIDs = $0.trackerIDs,
-                  let categoryName = $0.categoryName else { return nil}
+            guard let categoryName = $0.categoryName else { return nil}
             
-            return TrackerCategory(trackerIDs: TrackerCategory.trackerIDsFromString(udids: trackerIDs),
-                                categoryName: categoryName)
+            return TrackerCategory(categoryName: categoryName)
         }
     }
     
     func deleteCategory(_ categoryName: String) -> Bool {
         guard let categoryObject = getCategoryObject(categoryName) else { return false}
-        
+        if (categoryObject.tracker?.count ?? 0) > 0 {
+            return false
+        }
         context.delete(categoryObject)
+        
+        do {  try context.save() } catch { return false }
+        
+        //delegate?.didUpdate()
         return true
     }
     
-    private func getCategoryObject(_ category: String) -> TrackerCategoryCoreData? {
+    func getCategoryObject(_ category: String) -> TrackerCategoryCoreData? {
+        
         let request = NSFetchRequest<TrackerCategoryCoreData>(entityName: "TrackerCategoryCoreData")
         request.returnsObjectsAsFaults = false
         request.predicate = NSPredicate(format: "%K == %@", #keyPath(TrackerCategoryCoreData.categoryName), category)
         
         var records:[TrackerCategoryCoreData] = []
         do { records = try context.fetch(request) } catch { return nil }
+        
         return records.first
     }
-    
-    
 }
+
+// MARK: - NSFetchedResultsControllerDelegate
+extension TrackerCategoryStore: NSFetchedResultsControllerDelegate {
+    
+    func controller( _ controller: NSFetchedResultsController<NSFetchRequestResult>,
+                     didChange anObject: Any,
+                     at indexPath: IndexPath?,
+                     for type: NSFetchedResultsChangeType,
+                     newIndexPath: IndexPath?) {
+
+        self.delegate?.didUpdate()
+
+    }
+}
+
